added new features

+++ MyRecipe/main/views.py
@@ -1,7 +1,8 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required
-from .models import Recipe
-from .forms import RecipeForm
+from django.http import HttpResponseForbidden
+from .models import Recipe, Review
+from .forms import RecipeForm, ReviewForm
 
 def index(request):
     recipes = Recipe.objects.all()
@@ -9,12 +10,28 @@ def index(request):
 
 def recipe_detail(request, recipe_id):
     recipe = get_object_or_404(Recipe, pk=recipe_id)
-    return render(request, 'main/recipe_detail.html', {'recipe': recipe})
+    reviews = Review.objects.filter(recipe=recipe).order_by('-created_at')
 
+    if request.method == 'POST' and request.user.is_authenticated:
+        form = ReviewForm(request.POST)
+        if form.is_valid():
+            review = form.save(commit=False)
+            review.recipe = recipe
+            review.user = request.user
+            review.save()
+            return redirect('main:recipe_detail', recipe_id=recipe.id)
+    else:
+        form = ReviewForm()
+
+    return render(request, 'main/recipe_detail.html', {
+        'recipe': recipe,
+        'reviews': reviews,
+        'form': form
+    })
 @login_required
 def add_recipe(request):
     if request.method == 'POST':
-        form = RecipeForm(request.POST)
+        form = RecipeForm(request.POST, request.FILES)
         if form.is_valid():
             recipe = form.save(commit=False)
             recipe.created_by = request.user
@@ -27,18 +44,22 @@ def add_recipe(request):
 @login_required
 def update_recipe(request, recipe_id):
     recipe = get_object_or_404(Recipe, pk=recipe_id)
+    if recipe.created_by != request.user:
+        return HttpResponseForbidden("You do not have permission to edit this recipe.")
     if request.method == 'POST':
-        form = RecipeForm(request.POST, instance=recipe)
+        form = RecipeForm(request.POST, request.FILES, instance=recipe)
         if form.is_valid():
             form.save()
             return redirect('main:recipe_detail', recipe_id=recipe.id)
     else:
         form = RecipeForm(instance=recipe)
-    return render(request, 'main/update_recipe.html', {'form': form})
+    return render(request, 'main/update_recipe.html', {'form': form, 'recipe':recipe})
 
 @login_required
 def delete_recipe(request, recipe_id):
     recipe = get_object_or_404(Recipe, pk=recipe_id)
+    if recipe.created_by != request.user:
+        return HttpResponseForbidden("You do not have permission to delete this recipe.")
     if request.method == 'POST':
         recipe.delete()
         return redirect('main:index')
